+++ app/Providers/AuthServiceProvider.php
@@ -2,8 +2,12 @@
 
 namespace App\Providers;
 
+use App\Policies\RolePolicy;
 use App\Policies\RoomPolicy;
+use App\Policies\UserPolicy;
+use App\Role;
 use App\Room;
+use App\User;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\Gate;
 
@@ -17,6 +21,8 @@ class AuthServiceProvider extends ServiceProvider
     protected $policies = [
         // 'App\Model' => 'App\Policies\ModelPolicy',
         Room::class => RoomPolicy::class,
+        Role::class => RolePolicy::class,
+        User::class => UserPolicy::class,
     ];
 
     /**