+++ resources/js/app.js
@@ -38,12 +38,12 @@ if (process.env.NODE_ENV === 'development') {
  */
 Vue.config.errorHandler = function (error, vm, info) {
   const responseStatus = error.response !== undefined ? error.response.status : undefined;
-  const errorMessage = error.response.data ? error.response.data.message : undefined;
+  const errorMessage = error.response && error.response.data ? error.response.data.message : undefined;
 
   if (responseStatus === env.HTTP_UNAUTHORIZED) { // 401 => unauthorized, redirect and show error messages as flash!
     if (vm.$store.getters['session/isAuthenticated']) {
       vm.flashMessage.info(vm.$t('app.flash.unauthenticated'));
-      vm.$store.commit('session/setCurrentUser', null);
+      vm.$store.commit('setCurrentUser', { currentUser: null, emit: false });
       vm.$router.replace({ name: 'login' });
     }
   } else if (responseStatus === env.HTTP_FORBIDDEN && errorMessage === 'This action is unauthorized.') { // 403 => unauthorized, show error messages as flash!