+++ resources/js/lang/de/settings.js
@@ -31,7 +31,6 @@ export default {
     },
 
     default: 'Standard',
-    actions: 'Aktionen',
     nodata: 'Es sind keine Rollen vorhanden!',
 
     noOptions: 'Keine Berechtigungen vorhanden!',
@@ -44,7 +43,36 @@ export default {
   },
 
   users: {
-    title: 'Benutzer'
+    title: 'Benutzer',
+
+    new: 'Neuen Benutzer erstellen',
+    view: 'Detaillierte Informationen für den Benutzer {firstname} {lastname}',
+    edit: 'Benutzer {firstname} {lastname} bearbeiten',
+
+    nodata: 'Es sind keine Benutzer vorhanden!',
+    nodataFiltered: 'Für die Suchanfrage wurden keine Benutzer gefunden!',
+
+    id: 'ID',
+    firstname: 'Vorname',
+    lastname: 'Nachname',
+    email: 'E-Mail',
+    password: 'Passwort',
+    password_confirmation: 'Passwort bestätigen',
+    user_locale: 'Sprache',
+    roles: 'Rollen',
+    select_locale: 'Bitte wählen Sie eine Sprache aus',
+    removeRole: 'Rolle entfernen',
+    authenticator: {
+      title: 'Anmeldeart',
+      users: 'Registrierter Nutzer',
+      ldap: 'LDAP'
+    },
+
+    delete: {
+      item: 'Benutzer {firstname} {lastname} löschen',
+      confirm: 'Wollen Sie den Benutzer {firstname} {lastname} wirklich löschen?',
+      title: 'Benutzer löschen?'
+    }
   },
 
   application: {