+++ resources/js/lang/en/settings.js
@@ -30,7 +30,6 @@ export default {
       }
     },
     default: 'Default',
-    actions: 'Actions',
     nodata: 'No roles found!',
 
     noOptions: 'No permissions found!',
@@ -43,7 +42,36 @@ export default {
   },
 
   users: {
-    title: 'Users'
+    title: 'Users',
+
+    new: 'Create new user',
+    view: 'Detailed information for the user {firstname} {lastname}',
+    edit: 'Edit user {firstname} {lastname}',
+
+    nodata: 'No users found!',
+    nodataFiltered: 'For the filter query no users were found!',
+
+    id: 'ID',
+    firstname: 'Firstname',
+    lastname: 'Lastname',
+    email: 'E-Mail',
+    password: 'Password',
+    password_confirmation: 'Password confirmation',
+    user_locale: 'Language',
+    roles: 'Roles',
+    select_locale: 'Please select a language',
+    removeRole: 'Remove role',
+    authenticator: {
+      title: 'Authentication Type',
+      users: 'Registered User',
+      ldap: 'LDAP'
+    },
+
+    delete: {
+      item: 'Delete user {firstname} {lastname}',
+      confirm: 'Are you really want to delete the user {firstname} {lastname}?',
+      title: 'Delete user?'
+    }
   },
 
   application: {