+++ resources/js/store/modules/session.js
@@ -25,9 +25,9 @@ const actions = {
     await auth.login(credentials, method);
     await dispatch('getCurrentUser');
 
-    if (state.currentUser.locale !== null) {
-      await loadLanguageAsync(state.currentUser.locale);
-      commit('setCurrentLocale', state.currentUser.locale);
+    if (state.currentUser.user_locale !== null) {
+      await loadLanguageAsync(state.currentUser.user_locale);
+      commit('setCurrentLocale', state.currentUser.user_locale);
     }
   },
 
@@ -40,13 +40,13 @@ const actions = {
   async getCurrentUser ({ commit }) {
     let currentUser = await auth.getCurrentUser();
     if ($.isEmptyObject(currentUser)) { currentUser = null; }
-    commit('setCurrentUser', currentUser);
+    commit('setCurrentUser', { currentUser });
   },
 
   async logout ({ commit }) {
     commit('loading', null, { root: true });
     await auth.logout();
-    commit('setCurrentUser', null);
+    commit('setCurrentUser', { currentUser: null, emit: false });
     commit('loadingFinished', null, { root: true });
   },
 
@@ -66,9 +66,10 @@ const mutations = {
     state.settings = settings;
   },
 
-  setCurrentUser (state, currentUser) {
+  setCurrentUser (state, { currentUser, emit = true }) {
+    console.log('test', currentUser, emit);
     state.currentUser = currentUser;
-    PermissionService.setCurrentUser(state.currentUser);
+    PermissionService.setCurrentUser(state.currentUser, emit);
   }
 };