+++ routes/api.php
@@ -40,11 +40,8 @@ Route::prefix('v1')->namespace('api\v1')->name('api.v1.')->group(function () {
 
 // TODO: Implement or remove this completely
 //        Route::post('register', 'RegisterController@register');
-//
 //        Route::post('password/reset', 'ResetPasswordController@reset');
 //        Route::post('password/email', 'ForgotPasswordController@sendResetLinkEmail');
-//        Route::post('password/confirm', 'ConfirmPasswordController@confirm');
-//
 //        Route::post('email/resend', 'VerificationController@resend');
 //        Route::get('email/verify/{id}/{hash}', 'VerificationController@verify');
     });
@@ -80,6 +77,7 @@ Route::prefix('v1')->namespace('api\v1')->name('api.v1.')->group(function () {
         });
 
         Route::get('users/search','UserController@search')->name('users.search');
+        Route::apiResource('users', 'UserController');
 
         Route::get('roomTypes', 'RoomTypeController@index')->name('roomTypes.index');
     });
@@ -92,13 +90,7 @@ Route::prefix('v1')->namespace('api\v1')->name('api.v1.')->group(function () {
         Route::get('rooms/{room}/files/{file}', 'RoomFileController@show')->name('rooms.files.show')->middleware(['can:downloadFile,room,file', 'room.authenticate']);
     });
 
-
-
-
-
     Route::get('meetings/{meeting}/endCallback','MeetingController@endMeetingCallback')->name('meetings.endcallback');
-
-
 });
 
 if (!env('DISABLE_CATCHALL_ROUTES')) {